refactor(react): isolate media synchronization internals
